project: Fix various warnings

+++ source/nvidia/cv/nvidia-cv.cpp
@@ -88,7 +88,8 @@ streamfx::nvidia::cv::cv::cv()
 		env_size = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, nullptr, 0);
 		if (env_size > 0) {
 			buffer.resize(static_cast<size_t>(env_size) + 1);
-			env_size     = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, buffer.data(), buffer.size());
+			env_size     = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, buffer.data(),
+                                               static_cast<DWORD>(buffer.size()));
 			vfx_sdk_path = std::wstring(buffer.data(), buffer.size());
 		} else {
 			PWSTR   str = nullptr;
@@ -116,7 +117,8 @@ streamfx::nvidia::cv::cv::cv()
 		env_size = GetEnvironmentVariableW(ST_ENV_NVIDIA_AR_SDK_PATH, nullptr, 0);
 		if (env_size > 0) {
 			buffer.resize(static_cast<size_t>(env_size) + 1);
-			env_size    = GetEnvironmentVariableW(ST_ENV_NVIDIA_AR_SDK_PATH, buffer.data(), buffer.size());
+			env_size =
+				GetEnvironmentVariableW(ST_ENV_NVIDIA_AR_SDK_PATH, buffer.data(), static_cast<DWORD>(buffer.size()));
 			ar_sdk_path = std::wstring(buffer.data(), buffer.size());
 		} else {
 			PWSTR   str = nullptr;
